Create docker-entrypoint.sh
Python application often needs multiple Python files for different functions and purposes etc. During the cycle of development, it is essential to containerize the Python application and deploy the containers via Docker. One option to loop through all Python scripts is to introduce a Shell script as ENTRYPOINT in Dockerfile 

Things to keep in mind:
1) Make .sh executable before docker build command in terminal using chmod +x ~/path/to/.sh/file
2) Run regular docker build & docker run commands